# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_all-in-one_internal_http.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handler.createTokenRequest:
    properties:
      app:
        type: string
      name:
        type: string
      scope_prefix:
        type: string
    type: object
  handler.createTokenResponse:
    properties:
      app:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      notice:
        type: string
      revoked_at:
        type: string
      scope_prefix:
        type: string
      token:
        type: string
      token_prefix:
        type: string
    type: object
  model.AppToken:
    properties:
      app:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      revoked_at:
        type: string
      scope_prefix:
        type: string
      token_prefix:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: All-in-one is a multi-functional apps with the main goal for learning
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: All-in-one API
  version: "1.0"
paths:
...
  /ratelimit/tokens:
    get:
      description: List every app token (including revoked ones — the audit trail).
        The plaintext token and its hash are never returned (admin-only).
      produces:
      - application/json
      responses:
        "200":
          description: Tokens
          schema:
            allOf:
            - $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AppToken'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: List app tokens
      tags:
      - rate-limiting
    post:
      consumes:
      - application/json
      description: Mint a new app token for another app to call the rate-limit check
        API. The plaintext token is returned ONCE in this response and can never be
        retrieved again (admin-only).
      parameters:
      - description: App, name, optional scope prefix
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created token (plaintext shown once)
          schema:
            allOf:
            - $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.createTokenResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: Create an app token
      tags:
      - rate-limiting
  /ratelimit/tokens/{id}:
    delete:
      description: Soft-revoke an app token by id (the audit row is kept). Takes effect
        immediately (admin-only).
      parameters:
      - description: Token id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Revoked
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
        "403":
          description: Forbidden (not an admin)
          schema:
            $ref: '#/definitions/github_com_all-in-one_internal_http.Response'
      security:
      - BearerAuth: []
        DirectAuth: []
      summary: Revoke an app token
      tags:
      - rate-limiting
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  DirectAuth:
    description: Direct authentication bypass for testing (only works when DirectAuthEnabled
      is true in config). Provide username directly.
    in: header
    name: x-direct-auth-username
    type: apiKey
swagger: "2.0"